Refactoring: Move the functions likeUnlikePattern and getUserLikedPattern to the pattern service, because the promise expect the pattern interface, delete the UserToPattern interface, because its not required
